src/requirements.txt
django
mysqlclient
djangorestframework
django-filter
python-decouple